Taking advantage of Array shape
There were a lot of middlemen:
 - _iterator builds an array
 - .blocks was being accessed
 - to call _gen_ids
 - which built an array
 - to get the shape

I don't see the need of doing so many layers. The proposal cuts on several of those. Maybe it could even skip the first _iterator call, but that was the code that was there at the begining and maybe it is more future-proof for potential changes on the _iterator and/or array internals.